added registration page and functionality, and bug fixes, improvements

+++ CMS_TEMPLATE/admin/includes/view_all_posts.php
@@ -146,7 +146,7 @@ while($row = mysqli_fetch_assoc($view_all_posts_query)) {
     echo "<td>{$post_date}</td>";
     echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";
     echo "<td><a href='admin_posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-    echo "<td><a href='admin_posts.php?delete={$post_id}'>Delete</a></td>";
+    echo "<td><a onClick = \"javascript: return confirm('Are you sure you want to delete?'); \"href='admin_posts.php?delete={$post_id}'>Delete</a></td>";
     echo "</tr>";
 }
